Added visibility toggling of output variables
Data displayed on the 'Data Display' and 'DataPoint' windows can now have variables chosen to be hidden or not. There is no in-game interface option for this yet.
A new script 'OutputVariableVisibilty.cs' has been made to house the options and provide functions to change visibilities, and is a component of the 'DataFiles' gameobject.
'RocketAnimationUI.cs' and 'DataPoint'.cs has been updated to check for the visibility component and only display the chosen variables within it.
'DataPoint.cs' select method has been changed to update the text when the point is selected to reflect variable toggles, and now destroys the window on deselect instead of hiding it.