Added (Untested) Support for HTC Vive, Change Movement and Laser Controls
Movement is now full locomotion for the left joystick/ touchpad axes, and rotation linked to the right joystick horizontal axis.
This still has not been tested on a HTC Vive, but the new controls work with the Quest 2.
The laser pointer is now binded to the right controllers grip, which both the Quest 2 and Vive has. Before, if was binded to the joystick touch, which the Vive doesn't have.