Added axis variable selection to graph creator
The graph creator script will now accept different variables when creating a graph. These can be chosen in the graph creator objects inspector window, where there are fields for each axis and a list of acceptable axis names.
In the inspector, there is a context menu option for graph spawning, which will use the variables given in the inspector.
Added some more context menu options for the RocketAnimationUI.cs component, which allow for pausing and reseting of the animation, beyond just playing.