Added changing of axes on existing graphs
Existing graphs can now have their axis variables changed. This can be done in code, or in the inspector, with the Update Graph context menu option which will use the axes on the graph creator to update the latest graph.
When updating a graph, the graph is regenerated by deleting and recreating the datapoints in the right location. This should be changed to just move the datapoints to the new locations