Updated changing of axis variables on existing graphs + Performance improvements
Improved how the graphs were updated. Instead of deleting and remaking the datapoints, the points are now moved to new locations to represent the updated axes.
There is also an option to allow motion tweening for this movement, allowing the points to either teleport to where they need to be, or smoothly translate to the new locations.
The spheres used for the datapoints have also been updated, bringing down the tri count for each sphere from ~700 to 80, which also drops the total tris in the scene for one graph by about 90%