Added choosing of different input folders + Major fixes on freezing when creating/ updating a graph
The graph creator inspect now has a folder name input and a list of acceptable folder names for the user to choose, which will change which simulation files the graph will be based on.
The graphs no longer start at 0,0 (Update to CSVDataSource), since this affected x y data that had large negative numbers and small positive numbers.
Individual trajectories now get staggered between frames using coroutines to prevent the app from freezing for a period of time, now the changes are live (however at a lower framerate). This has been applied to both creation and updating of graphs.
The graph config now has an option to move the points as they are updating during the coroutines, or to wait for all trajectories to update before moving them in bulk.