Maximize function written.
The maximizer object contains a panel, that will need to be positioned to suit the panel being minimized, and a button that is already programmed to listen for clicks.
To add other panels to the code, the panel will need to be dragged from the hierarchy into the inspector where  "Other Panel" is listed.
A button will also need to be added to the original panel. The maximise panel will disappear at runtime.