Updated graph spawning and deleting
Added a dedicated component to handle graph positioning, which is now attached to the graph creator.
All previous code for this was moved over to it, and more functionality was added.
The graphs are now tracked and prevent too many graphs being made. When a graph is deleted, the layout moves and the gap is filled by other graphs.