Attached Graph Creator to hand, various fixes made
The Graph Creator menu is now attached to the users left hand, allowing them to interact with the creator menu where ever they are in the scene. They can also hide and show the menu, so that it doesn't always take up visual space on the hand.
Lots of small bug fixes were made, primarily with the height and rotations of spawned graphs. This has been tested in VR, and the graphs now spawn at an appropriate height, and the bar graph prefab has been modified to line up and rotate properly as well.