Added VRTK Canvas to gConfigs, Fixed Rocket Issue
A VRTK canvas was added to the graph config interfaces in the prefab so that it can be interacted with in VR.
Testing the graph config in VR brought up an issue where the rocket wasn't attaching to the proper trajectories. The rocket now gets deactivated until trajectories are finished to be placed again.