Updated graph spawning
The movement feature in the graph handler has been removed, as with the area selection mechanic, there wasn't really a good way to keep it in without disregarding the players choice when selecting spots.
The way graphs are stored is now slightly different, instead of a list only containing created graphs, all blank spots are filled with nulls to keep the offsets easily.