Added changing of inputs on existing bar graphs
Bar graphs can now have their input variables changed after creation, sort of.
To the user visually, the bar graph appears to be updated properly, however it gets deleted and spawned within the frame this is applied, effectively doing the same thing.
This was done as it is far, far more simpler than modifying an already made graph.