Made all menus moveable
Made all menus moveable by linking interactable cubes to each menu. These can be found in the graph prefabs Movers game object. Both scatter and bar graphs are updated with this.
Moving cubes have been given new materials to match the interfaces. A bug that I've noticed with the outline shader I added is that it only renders in the left eye, which is slightly jarring, but not too major.