Updated BarGraph config
Added a dropdown bar to the menu.
Connected a set of folder names to the dropdown. Created a test method for desktop mode.
Connected the update graph button to a new method that uses gConfig's updatebargraph method.
Set the method to public and added it to the Context Menu for testing.