Fixed HMD offset, text container, and selected var warnings
Many of the warnings have been fixed.
The offset warning was fixed by changing the camera offset to default instead of floor.
The text containers were removed from the iatk axis prefab and the legend to fix the warnings.
The selected parameter not existing warning was fixed by removing the slider animation transition on the rocket ui, and removing the animator component