Fixed all runtime warnings, fixed legend overlapping new items
All warnings during runtime is now fixed, with the program experience no warnings when being used. The only case there should be a warning is when it detects theres no HMD detected, however this seems to be random, and only happen once.
The legend items now deletes old items when the input folder is changed, so the new ones spawned won't duplicate and overlap the old ones.