Fixed bars from being deleted by the mouse cursor
An artifact from the previous attempt to prevent the mouse from interfering with the bar graph when in VR mode. The bars would be deleted by the cursor if the mouse hovers over then leaves the bar when in game mode.
This has been fixed by preventing the trigger, hover, selection etc. events by the cursor from happening.